chore: refactor respondent block into smaller components
