+++ frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/RespondentBlock/Components/ConditionalRoutingOption.tsx
@@ -1,16 +1,14 @@
 import { useEffect, useMemo, useState } from 'react'
-import { Controller, useForm, UseFormReturn } from 'react-hook-form'
+import { Controller, useForm } from 'react-hook-form'
 import { BiPlus } from 'react-icons/bi'
 import { useParams } from 'react-router'
 import {
-  As,
   Button,
   FormControl,
   Stack,
   Text,
   useDisclosure,
 } from '@chakra-ui/react'
-import { get } from 'lodash'
 import Papa from 'papaparse'
 import isEmail from 'validator/lib/isEmail'
 
@@ -20,207 +18,24 @@ import {
   CONDITIONAL_ROUTING_INVALID_CSV_FORMAT_ERROR_MESSAGE,
   CONDITIONAL_ROUTING_MISMATCHED_OPTIONS_ERROR_MESSAGE,
 } from '~shared/constants/errors'
-import {
-  DropdownFieldBase,
-  FormFieldDto,
-  UserDto,
-  WorkflowType,
-} from '~shared/types'
+import { DropdownFieldBase, FormFieldDto, WorkflowType } from '~shared/types'
 import { checkIsOptionsMismatched } from '~shared/utils/options-recipients-map-validation'
 
-import { textStyles } from '~theme/textStyles'
 import { parseCsvFileToCsvString } from '~utils/parseCsvFileToCsvString'
 import { SingleSelect } from '~components/Dropdown'
 import Attachment from '~components/Field/Attachment'
 import { downloadFile } from '~components/Field/Attachment/utils/downloadFile'
 import FormErrorMessage from '~components/FormControl/FormErrorMessage'
-import FormLabel from '~components/FormControl/FormLabel'
 import Radio from '~components/Radio'
-import { TagInput } from '~components/TagInput'
 
 import { useEditFormField } from '~features/admin-form/create/builder-and-design/mutations/useEditFormField'
 import { BASICFIELD_TO_DRAWER_META } from '~features/admin-form/create/constants'
-import { EditStepInputs } from '~features/admin-form/create/workflow/types'
 import { FormFieldWithQuestionNo } from '~features/form/types'
-// TODO (MRF-Conditional-Routing): Remove useUser import when conditional routing is out of beta
-import { useUser } from '~features/user/queries'
 
-import { useAdminFormWorkflow } from '../../../hooks/useAdminFormWorkflow'
-import { isFirstStepByStepNumber } from '../utils/isFirstStepByStepNumber'
-
+import { WORKFLOW_TYPE_VALIDATION } from './common'
 import { ConditionalRoutingMappingDeleteModal } from './ConditionalRoutingMappingDeleteModal'
 import { ConditionalRoutingOptionModal } from './ConditionalRoutingOptionModal'
-import { EditStepBlockContainer } from './EditStepBlockContainer'
-
-const WORKFLOW_TYPE_VALIDATION = {
-  required: 'Please select a respondent type',
-  validate: (value: WorkflowType) => {
-    if (!Object.values(WorkflowType).includes(value)) {
-      return 'The selected respondent type is invalid'
-    }
-  },
-}
-
-interface RespondentOptionProps {
-  isLoading: boolean
-  formMethods: UseFormReturn<EditStepInputs>
-  selectedWorkflowType: WorkflowType
-}
-
-export interface FieldItem {
-  label: string
-  value: string
-  icon?: As
-}
-
-const StaticRespondentOption = ({
-  isLoading,
-  formMethods,
-  selectedWorkflowType,
-}: RespondentOptionProps) => {
-  const {
-    register,
-    control,
-    formState: { errors },
-  } = formMethods
-  const staticTagInputErrorMessage = get(errors, 'emails.message')
-
-  return (
-    <>
-      <Radio
-        isDisabled={isLoading}
-        isLabelFullWidth
-        allowDeselect={false}
-        value={WorkflowType.Static}
-        {...register('workflow_type', WORKFLOW_TYPE_VALIDATION)}
-        px="0.5rem"
-        __css={{
-          _focusWithin: {
-            boxShadow: 'none',
-          },
-        }}
-      >
-        <Text>Specific email(s)</Text>
-        {selectedWorkflowType === WorkflowType.Static ? (
-          <FormControl
-            pt="0.5rem"
-            isReadOnly={isLoading}
-            id="emails"
-            isRequired
-            isInvalid={staticTagInputErrorMessage}
-            key="emails"
-          >
-            <Controller
-              name="emails"
-              control={control}
-              rules={{
-                validate: {
-                  required: (emails) =>
-                    !emails || emails.length === 0
-                      ? 'You must enter at least one email to receive responses'
-                      : true,
-                  isEmails: (emails) =>
-                    !emails ||
-                    emails.every((email) => isEmail(email)) ||
-                    'Please enter valid email(s) (e.g. me@example.com) separated by commas, as invalid emails will not be saved',
-                },
-              }}
-              render={({ field }) => (
-                <TagInput
-                  isDisabled={isLoading}
-                  placeholder="me@example.com"
-                  tagValidation={isEmail}
-                  {...field}
-                />
-              )}
-            />
-            <FormErrorMessage>{staticTagInputErrorMessage}</FormErrorMessage>
-            {!staticTagInputErrorMessage ? (
-              <Text textStyle="body-2" color="secondary.400" mt="0.5rem">
-                Separate multiple emails with a comma
-              </Text>
-            ) : null}
-          </FormControl>
-        ) : null}
-      </Radio>
-    </>
-  )
-}
-
-interface DynamicRespondentOptionProps extends RespondentOptionProps {
-  emailFieldItems: FieldItem[]
-}
-
-const DynamicRespondentOption = ({
-  isLoading,
-  selectedWorkflowType,
-  formMethods,
-  emailFieldItems,
-}: DynamicRespondentOptionProps) => {
-  const {
-    register,
-    formState: { errors },
-    control,
-  } = formMethods
-
-  return (
-    <>
-      <Radio
-        isDisabled={isLoading}
-        isLabelFullWidth
-        allowDeselect={false}
-        value={WorkflowType.Dynamic}
-        {...register('workflow_type', WORKFLOW_TYPE_VALIDATION)}
-        px="0.5rem"
-        __css={{
-          _focusWithin: {
-            boxShadow: 'none',
-          },
-        }}
-      >
-        <Text>An email field from the form</Text>
-        {selectedWorkflowType === WorkflowType.Dynamic ? (
-          <FormControl
-            pt="0.5rem"
-            isReadOnly={isLoading}
-            id="field"
-            isRequired
-            isInvalid={!!errors.field}
-          >
-            <Controller
-              control={control}
-              name="field"
-              rules={{
-                required: 'Please select a field',
-                validate: (selectedValue) => {
-                  return (
-                    isLoading ||
-                    !emailFieldItems ||
-                    emailFieldItems.some(
-                      ({ value: fieldValue }) => fieldValue === selectedValue,
-                    ) ||
-                    'Field is not an email field'
-                  )
-                },
-              }}
-              render={({ field: { value = '', ...rest } }) => (
-                <SingleSelect
-                  isDisabled={isLoading}
-                  isClearable={false}
-                  placeholder="Select a field"
-                  items={emailFieldItems}
-                  value={value}
-                  {...rest}
-                />
-              )}
-            />
-            <FormErrorMessage>{errors.field?.message}</FormErrorMessage>
-          </FormControl>
-        ) : null}
-      </Radio>
-    </>
-  )
-}
+import { RespondentOptionProps } from './types'
 
 interface ConditionalRoutingOptionProps extends RespondentOptionProps {
   conditionalFormFields: FormFieldWithQuestionNo<
@@ -232,7 +47,7 @@ export interface ConditionalRoutingConfig {
   csvFile: File | null
 }
 
-const ConditionalRoutingOption = ({
+export const ConditionalRoutingOption = ({
   isLoading,
   formMethods,
   selectedWorkflowType,
@@ -633,89 +448,5 @@ const ConditionalRoutingOption = ({
         ) : null}
       </Radio>
     </>
-  )
-}
-
-interface RespondentBlockProps {
-  stepNumber: number
-  isLoading: boolean
-  formMethods: UseFormReturn<EditStepInputs>
-  user: UserDto | undefined
-}
-
-export const RespondentBlock = ({
-  stepNumber,
-  isLoading,
-  formMethods,
-}: RespondentBlockProps): JSX.Element => {
-  const {
-    formState: { errors },
-    watch,
-  } = formMethods
-
-  // TODO (MRF-Conditional-Routing): Remove isTest and user/useUser when conditional routing is out of beta
-  const { user } = useUser()
-  const isTest = import.meta.env.STORYBOOK_NODE_ENV === 'test'
-
-  const { emailFormFields = [], dropdownFormFields = [] } =
-    useAdminFormWorkflow()
-
-  const emailFieldItems = emailFormFields.map(
-    ({ _id, questionNumber, title, fieldType }) => ({
-      label: `${questionNumber}. ${title}`,
-      value: _id,
-      icon: BASICFIELD_TO_DRAWER_META[fieldType].icon,
-    }),
-  )
-
-  const selectedWorkflowType = watch('workflow_type')
-
-  const isFirstStep = isFirstStepByStepNumber(stepNumber)
-
-  return (
-    <EditStepBlockContainer>
-      {isFirstStep ? (
-        <Stack spacing="0.5rem">
-          <Text style={textStyles.h4}>Respondent in this step</Text>
-          <Text>Anyone who has access to your form</Text>
-        </Stack>
-      ) : (
-        <FormControl
-          isReadOnly={isLoading}
-          isRequired
-          isInvalid={!!errors.workflow_type}
-        >
-          <FormLabel style={textStyles.h4}>Select a respondent</FormLabel>
-          <Stack spacing="0.25rem">
-            <Radio.RadioGroup value={selectedWorkflowType}>
-              <DynamicRespondentOption
-                selectedWorkflowType={selectedWorkflowType}
-                emailFieldItems={emailFieldItems}
-                formMethods={formMethods}
-                isLoading={isLoading}
-              />
-              <StaticRespondentOption
-                selectedWorkflowType={selectedWorkflowType}
-                formMethods={formMethods}
-                isLoading={isLoading}
-              />
-              {/* TODO (MRF-Conditional-Routing): Remove isTest and user check when
-              conditional routing is out of beta */}
-              {isTest || user?.betaFlags?.mrfConditionalRouting ? (
-                <>
-                  <ConditionalRoutingOption
-                    selectedWorkflowType={selectedWorkflowType}
-                    conditionalFormFields={dropdownFormFields}
-                    formMethods={formMethods}
-                    isLoading={isLoading}
-                  />
-                </>
-              ) : null}
-            </Radio.RadioGroup>
-          </Stack>
-          <FormErrorMessage>{errors.workflow_type?.message}</FormErrorMessage>
-        </FormControl>
-      )}
-    </EditStepBlockContainer>
   )
 }
