+++ frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/RespondentBlock/Components/ConditionalRoutingOptionModal.stories.tsx
@@ -3,11 +3,11 @@ import { expect, screen, userEvent, waitFor } from '@storybook/test'
 
 import { StoryRouter } from '~utils/storybook'
 
+import { ConditionalRoutingConfig } from './ConditionalRoutingOption'
 import {
   ConditionalRoutingOptionModal,
   ConditionalRoutingOptionModalProps,
 } from './ConditionalRoutingOptionModal'
-import { ConditionalRoutingConfig } from './RespondentBlock'
 
 const ModalContainer = ({
   isOpen,
