+++ frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/RespondentBlock/Components/DynamicRespondentOption.tsx
@@ -1,0 +1,86 @@
+import { Controller } from 'react-hook-form'
+import { FormControl, Text } from '@chakra-ui/react'
+
+import { WorkflowType } from '~shared/types'
+
+import { SingleSelect } from '~components/Dropdown'
+import FormErrorMessage from '~components/FormControl/FormErrorMessage'
+import Radio from '~components/Radio'
+
+import { WORKFLOW_TYPE_VALIDATION } from './common'
+import { FieldItem, RespondentOptionProps } from './types'
+
+interface DynamicRespondentOptionProps extends RespondentOptionProps {
+  emailFieldItems: FieldItem[]
+}
+
+export const DynamicRespondentOption = ({
+  isLoading,
+  selectedWorkflowType,
+  formMethods,
+  emailFieldItems,
+}: DynamicRespondentOptionProps) => {
+  const {
+    register,
+    formState: { errors },
+    control,
+  } = formMethods
+
+  return (
+    <>
+      <Radio
+        isDisabled={isLoading}
+        isLabelFullWidth
+        allowDeselect={false}
+        value={WorkflowType.Dynamic}
+        {...register('workflow_type', WORKFLOW_TYPE_VALIDATION)}
+        px="0.5rem"
+        __css={{
+          _focusWithin: {
+            boxShadow: 'none',
+          },
+        }}
+      >
+        <Text>An email field from the form</Text>
+        {selectedWorkflowType === WorkflowType.Dynamic ? (
+          <FormControl
+            pt="0.5rem"
+            isReadOnly={isLoading}
+            id="field"
+            isRequired
+            isInvalid={!!errors.field}
+          >
+            <Controller
+              control={control}
+              name="field"
+              rules={{
+                required: 'Please select a field',
+                validate: (selectedValue) => {
+                  return (
+                    isLoading ||
+                    !emailFieldItems ||
+                    emailFieldItems.some(
+                      ({ value: fieldValue }) => fieldValue === selectedValue,
+                    ) ||
+                    'Field is not an email field'
+                  )
+                },
+              }}
+              render={({ field: { value = '', ...rest } }) => (
+                <SingleSelect
+                  isDisabled={isLoading}
+                  isClearable={false}
+                  placeholder="Select a field"
+                  items={emailFieldItems}
+                  value={value}
+                  {...rest}
+                />
+              )}
+            />
+            <FormErrorMessage>{errors.field?.message}</FormErrorMessage>
+          </FormControl>
+        ) : null}
+      </Radio>
+    </>
+  )
+}
