+++ frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/RespondentBlock/Components/StaticRespondentOption.tsx
@@ -1,0 +1,88 @@
+import { Controller } from 'react-hook-form'
+import { FormControl, Text } from '@chakra-ui/react'
+import { get } from 'lodash'
+import isEmail from 'validator/lib/isEmail'
+
+import { WorkflowType } from '~shared/types'
+
+import FormErrorMessage from '~components/FormControl/FormErrorMessage'
+import Radio from '~components/Radio'
+import { TagInput } from '~components/TagInput'
+
+import { RespondentOptionProps } from '../RespondentBlock'
+
+import { WORKFLOW_TYPE_VALIDATION } from './common'
+
+export const StaticRespondentOption = ({
+  isLoading,
+  formMethods,
+  selectedWorkflowType,
+}: RespondentOptionProps) => {
+  const {
+    register,
+    control,
+    formState: { errors },
+  } = formMethods
+  const staticTagInputErrorMessage = get(errors, 'emails.message')
+
+  return (
+    <>
+      <Radio
+        isDisabled={isLoading}
+        isLabelFullWidth
+        allowDeselect={false}
+        value={WorkflowType.Static}
+        {...register('workflow_type', WORKFLOW_TYPE_VALIDATION)}
+        px="0.5rem"
+        __css={{
+          _focusWithin: {
+            boxShadow: 'none',
+          },
+        }}
+      >
+        <Text>Specific email(s)</Text>
+        {selectedWorkflowType === WorkflowType.Static ? (
+          <FormControl
+            pt="0.5rem"
+            isReadOnly={isLoading}
+            id="emails"
+            isRequired
+            isInvalid={staticTagInputErrorMessage}
+            key="emails"
+          >
+            <Controller
+              name="emails"
+              control={control}
+              rules={{
+                validate: {
+                  required: (emails) =>
+                    !emails || emails.length === 0
+                      ? 'You must enter at least one email to receive responses'
+                      : true,
+                  isEmails: (emails) =>
+                    !emails ||
+                    emails.every((email) => isEmail(email)) ||
+                    'Please enter valid email(s) (e.g. me@example.com) separated by commas, as invalid emails will not be saved',
+                },
+              }}
+              render={({ field }) => (
+                <TagInput
+                  isDisabled={isLoading}
+                  placeholder="me@example.com"
+                  tagValidation={isEmail}
+                  {...field}
+                />
+              )}
+            />
+            <FormErrorMessage>{staticTagInputErrorMessage}</FormErrorMessage>
+            {!staticTagInputErrorMessage ? (
+              <Text textStyle="body-2" color="secondary.400" mt="0.5rem">
+                Separate multiple emails with a comma
+              </Text>
+            ) : null}
+          </FormControl>
+        ) : null}
+      </Radio>
+    </>
+  )
+}
