+++ frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/RespondentBlock/Components/common.ts
@@ -1,0 +1,10 @@
+import { WorkflowType } from '~shared/types/form/workflow'
+
+export const WORKFLOW_TYPE_VALIDATION = {
+  required: 'Please select a respondent type',
+  validate: (value: WorkflowType) => {
+    if (!Object.values(WorkflowType).includes(value)) {
+      return 'The selected respondent type is invalid'
+    }
+  },
+}
