+++ frontend/src/features/admin-form/create/workflow/components/WorkflowContent/EditStepBlock/RespondentBlock/index.tsx
@@ -1,0 +1,1 @@
+export { RespondentBlock } from './RespondentBlock'
